Add mod instruction to valid instruction list and instruction decoder
The instruction list provides a bit pattern that can be
pattern-matched against using Chisel3's framework for handling bit
sequences, which will generate the proper comparators needed in
Verilog, and therefore the proper hardware.

The next part, the decoder, matches against the instruction list, and
constructs a list of signals that are passed throughout the CPU, as
the control signals.

The modulo instruction is intended for illustration, and reuses ALL
components of the remainder unit from division. This means its
semantics are not _technically_ correct, as modulo of negative numbers
behaves differently than remainders.