rocket-chip test generator will use modulo ISA test
This test *must* be named the same thing as the .h file in the
riscv-tests repository!
It is probably best if this test has the same name as the instruction
itself too, though I am not sure how true that is.

This Scala code generates a "Makefile" (in quotes because it
is *sourced* by an actual Makefile and has the syntax of a Makefile,
but it *cannot* be run on its own). This generated Makefile includes
everything needed to run not only the ISA tests, but also the
performance evaluation tests on the generated design.