Use standards-removed UIE to determine if pipelined deleg enabled
For now, I can use the UIE bit in the mstatus register as a proxy to
determine if pipelined exception/interrupt delegation is enabled.

In the long-run, it would be better to create either another CSR or
codify this mstatus bit as actually determining whether or not
pipelined exception/interrupt delegation is enabled.