Add CSR to inform user-handler of double-fault state
It is still up to SW to figure out what to do if we actually hit a
double-fault case. Currently the HW provides a way to detect such an
event.

In the future, it would probably make sense for a double-fault to
re-raise the exception into supervisor mode for the kernel to handle.
This would have to set the various supervisor-level CSRs with the
values that we just populated into the user-level CSRs (namely sepc,
spp, scause).