Fix recursive type issue with FP exception signal in writeback stage
In the previous commit, the Scala compiler returns "scala recursive
value wb_xcpt needs type" if csr.io.fp_xcpt is present in the
writeback exception & cause sequence.

After some investigation, this kind of error is a semi-red herring.
What it really means is that Scala either could not figure out the
type of the thing, or that a nother symbol with the same identifier
was introduced by an intermediate pass that causes some other kinds of
issues.

[1] https://stackoverflow.com/a/44211671
[2] https://users.scala-lang.org/t/recursive-value-needs-type/5246/11