Use Rocket's hidden internal current privilege mode to control U/SRET
In the case where you go U->U->S, you end up in a state where upon
entering S-mode you have:
  * UIE = 1
  * reg_ualready_handling = true.B
  * SPP = U
  * SIE = 1

This set of nested exceptions is easy to reach. For example, a
printf() in a pipelined exception handler.

This USED to mean that upon sret out of S-mode, UEPC would be used as
the return PC, rather than SEPC.

The goal to produce a system that behaves "Always NO change when in in
privilege when IN User, but changes privilege upon switching to S".